# [quantization] Copy quantize routine to vec256