@raghuramank100 2019-09-05
Since the change only pertains to quantize_linear, can you benchmark the speed improvement for just that part so that we dont have the overhead of add?