@jamesr66a 2019-09-05
@raghuramank100 this change does not affect quantize_linear